feat: Migrate to running gluon-lang.org in ECS
Currently there is just lonely EC2 instance + cron script which tries to
keep the server up and running. This isn't great as I need to ssh in to
look at logs and the service only restarts once per day.

By moving to ECS we get logs in cloudwatch, automatic restarts on
crashes and we can setup alarms for any error cases. (Terraform is
certainly overkill for this, but setting up a new terraform environment
is good for learning, I already have a decent understanding of maintaining
one so).

cc https://github.com/gluon-lang/gluon/issues/881